fix(e2e/epp): treat non-HTTP-200 as failure in generateTraffic; split test helpers
