+++ test/e2e/epp/e2e_test.go
@@ -17,16 +17,10 @@ limitations under the License.
 package epp
 
 import (
-	"context"
-	"encoding/json"
 	"errors"
 	"fmt"
 	"strconv"
-	"strings"
-	"sync"
-	"time"
 
-	"github.com/google/go-cmp/cmp/cmpopts"
 	"github.com/onsi/ginkgo/v2"
 	"github.com/onsi/gomega"
 	appsv1 "k8s.io/api/apps/v1"
@@ -37,20 +31,10 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	v1 "sigs.k8s.io/gateway-api-inference-extension/api/v1"
 	"sigs.k8s.io/gateway-api-inference-extension/apix/v1alpha2"
-	"sigs.k8s.io/gateway-api-inference-extension/pkg/epp/metadata"
 
 	testutils "github.com/llm-d/llm-d-inference-scheduler/test/utils/igw"
 )
 
-const (
-	firstPort             = 8000
-	numPorts              = 2
-	maxConcurrentRequests = 2 // prevent hammering Envoy and backend
-	maxRetries            = 5
-	backoff               = 5 * time.Second
-	batches               = 20
-	apiEmbeddings         = "/embeddings"
-)
 
 var _ = ginkgo.Describe("InferencePool", func() {
 	var infObjective *v1alpha2.InferenceObjective
@@ -282,496 +266,3 @@ var _ = ginkgo.Describe("InferencePool", func() {
 		})
 	})
 })
-
-// newInferenceObjective creates an InferenceObjective in the given namespace for testutils.
-func newInferenceObjective(ns string) *v1alpha2.InferenceObjective {
-	return testutils.MakeModelWrapper(types.NamespacedName{Name: "inferenceobjective-sample", Namespace: ns}).
-		SetPriority(2).
-		SetPoolRef(modelServerName).
-		Obj()
-}
-
-// verifyTrafficRouting contains the logic for the "Should route traffic to target model servers" test.
-func verifyTrafficRouting() {
-	ginkgo.By("Verifying traffic routing")
-	for _, t := range []struct {
-		api              string
-		promptOrMessages any
-	}{
-		{
-			api:              "/completions",
-			promptOrMessages: "Write as if you were a critic: San Francisco",
-		},
-		{
-			api: "/chat/completions",
-			promptOrMessages: []map[string]any{
-				{
-					"role":    "user",
-					"content": "Write as if you were a critic: San Francisco",
-				},
-			},
-		},
-		{
-			api: "/chat/completions",
-			promptOrMessages: []map[string]any{
-				{
-					"role":    "user",
-					"content": "Write as if you were a critic: San Francisco",
-				},
-				{"role": "assistant", "content": "Okay, let's see..."},
-				{"role": "user", "content": "Now summarize your thoughts."},
-			},
-		},
-		{
-			api:              apiEmbeddings,
-			promptOrMessages: "The food was delicious and the service was great.",
-		},
-		{
-			api:              apiEmbeddings,
-			promptOrMessages: []string{"First sentence to embed.", "Second sentence to embed."},
-		},
-	} {
-		ginkgo.By(fmt.Sprintf("Verifying connectivity through the inference extension with %s api and prompt/messages: %v", t.api, t.promptOrMessages))
-
-		// Skip embeddings API if server returns 404 (not all models support embeddings).
-		if t.api == apiEmbeddings {
-			probeCmd := getCurlCommand(envoyName, testConfig.NsName, envoyPort, modelName, curlTimeout, t.api, t.promptOrMessages, false)
-			probeResp, probeErr := testutils.ExecCommandInPod(testConfig, "curl", "curl", probeCmd)
-			if probeErr == nil && strings.Contains(probeResp, "404") {
-				ginkgo.Skip("Skipping " + apiEmbeddings + ": server returned 404 (embeddings may not be supported by this model)")
-			}
-		}
-
-		// Expected ports and client-facing model name (response model is rewritten back to the incoming name)
-		expectedPort := generateSequence(firstPort, numPorts)
-		expectedModel := []string{modelName}
-
-		// Observed ports and InferenceObjective target models
-		actualModel := make(map[string]int)
-		actualPort := make(map[int]int)
-
-		// Send curl requests to verify routing to all target ports in the InferencePool.
-		// Run a small batch per retry (e.g., 5) to keep the test active
-		for i := range batches {
-			uniqueID := time.Now().UnixNano()
-			dynamicHashValue := fmt.Sprintf("Nonce-%d", uniqueID)
-			currentPromptOrMessages := t.promptOrMessages // Start with the original
-
-			// Check if the payload is a slice of maps (e.g., for /chat/completions)
-			if originalMessages, ok := currentPromptOrMessages.([]map[string]any); ok {
-				nonceMsg := map[string]any{
-					"role":    "system",
-					"content": fmt.Sprintf("TestNonce: %s-%d", dynamicHashValue, i),
-				}
-
-				currentPromptOrMessages = append([]map[string]any{nonceMsg}, originalMessages...)
-			} else if originalString, ok := t.promptOrMessages.(string); ok {
-				currentPromptOrMessages = fmt.Sprintf("[TestNonce: %s-%d] %s", dynamicHashValue, i, originalString)
-			} else if originalStrings, ok := t.promptOrMessages.([]string); ok {
-				// For embeddings with array input, prepend a unique string so each request is distinct.
-				withNonce := make([]string, 0, len(originalStrings)+1)
-				withNonce = append(withNonce, fmt.Sprintf("[TestNonce: %s-%d]", dynamicHashValue, i))
-				withNonce = append(withNonce, originalStrings...)
-				currentPromptOrMessages = withNonce
-			} else {
-				currentPromptOrMessages = t.promptOrMessages
-			}
-
-			curlCmd := getCurlCommand(envoyName, testConfig.NsName, envoyPort, modelName, curlTimeout, t.api, currentPromptOrMessages, false)
-
-			var resp string
-			var err error
-			// Repeatedly send a message until we get a successful response.
-			for attempt := 0; attempt <= maxRetries; attempt++ {
-				resp, err = testutils.ExecCommandInPod(testConfig, "curl", "curl", curlCmd)
-				if err == nil && strings.Contains(resp, "200 OK") {
-					break // Success!
-				}
-
-				if attempt < maxRetries {
-					time.Sleep(backoff)
-				}
-			}
-
-			gomega.Expect(err).ToNot(gomega.HaveOccurred(), "Expected curl command to succeed")
-			gomega.Expect(resp).To(gomega.ContainSubstring("200 OK"), "Expected to receive a 200 OK response...")
-
-			for _, m := range expectedModel {
-				if strings.Contains(resp, m) {
-					actualModel[m] = 0
-				}
-			}
-			for _, p := range expectedPort {
-				if strings.Contains(resp, fmt.Sprintf("x-inference-port: %d", p)) {
-					actualPort[p] = 0
-				}
-			}
-		}
-
-		gotModel := make([]string, 0, len(actualModel))
-		for m := range actualModel {
-			gotModel = append(gotModel, m)
-		}
-		gotPort := make([]int, 0, len(actualPort))
-		for p := range actualPort {
-			gotPort = append(gotPort, p)
-		}
-
-		ginkgo.GinkgoWriter.Printf("Port distribution: %v\n", actualPort)
-		ginkgo.GinkgoWriter.Printf("Model distribution: %v\n", actualModel)
-
-		gomega.Expect(gotModel).To(gomega.BeComparableTo(expectedModel, cmpopts.SortSlices(func(a, b string) bool { return a < b })))
-		gomega.Expect(gotPort).To(gomega.BeComparableTo(expectedPort, cmpopts.SortSlices(func(a, b int) bool { return a < b })))
-	}
-}
-
-// verifyMetrics contains the logic for the "Should expose EPP metrics after generating traffic" test.
-func verifyMetrics() {
-	ginkgo.By("Verifying metrics exposure")
-
-	// Generate traffic by sending requests through the inference extension.
-	ginkgo.By("Generating traffic through the inference extension")
-	curlCmd := getCurlCommand(envoyName, testConfig.NsName, envoyPort, modelName, curlTimeout, "/completions", "Write as if you were a critic: San Francisco", true)
-
-	// Run the curl command multiple times to generate some metrics data.
-
-	semaphore := make(chan struct{}, maxConcurrentRequests)
-
-	errorGood := generateTraffic(curlCmd, batches, semaphore)
-	gomega.Expect(errorGood).NotTo(gomega.HaveOccurred(), "Expected good traffic generation to succeed")
-
-	// Modify the curl command to generate some error metrics.
-	curlCmd[len(curlCmd)-1] = "invalid input"
-	errorBad := generateTraffic(curlCmd, batches, semaphore)
-	gomega.Expect(errorBad).NotTo(gomega.HaveOccurred(), "Expected bad traffic generation to succeed")
-
-	// looks like a flaky test, will investigate separately
-	ginkgo.By("Verifying that all expected metrics are present.")
-	ginkgo.Skip("Skipping flaky metrics verification test - will investigate separately")
-
-	// Now scrape metrics from the EPP endpoint via the curl pod.
-	ginkgo.By("Scraping metrics from the EPP endpoint and verifying all backends were hit")
-	podIP := findReadyPod().Status.PodIP
-
-	// Get the authorization token for reading metrics.
-	token := ""
-	gomega.Eventually(func(g gomega.Gomega) {
-		t, err := getMetricsReaderToken(testConfig.K8sClient)
-		g.Expect(err).NotTo(gomega.HaveOccurred())
-		g.Expect(t).NotTo(gomega.BeEmpty())
-		token = t
-	}, testConfig.ExistsTimeout, testConfig.Interval).Should(gomega.Succeed())
-
-	// Construct the metric scraping curl command using Pod IP.
-	metricScrapeCmd := getMetricsScrapeCommand(podIP, token)
-
-	modelServerPods, err := getPodsByLabel(testConfig.Context, testConfig.K8sClient, testConfig.NsName, "app", modelServerName)
-	gomega.Expect(err).NotTo(gomega.HaveOccurred(), "Expected to find model server pods")
-
-	// Define the metrics we expect to see
-	preset := []string{ //nolint:prealloc
-		"inference_objective_request_total",
-		"inference_objective_request_error_total",
-		"inference_objective_request_duration_seconds",
-		"inference_objective_normalized_time_per_output_token_seconds",
-		"inference_objective_request_sizes",
-		"inference_objective_response_sizes",
-		"inference_objective_input_tokens",
-		"inference_objective_output_tokens",
-		"inference_pool_average_kv_cache_utilization",
-		"inference_pool_average_queue_size",
-		"inference_pool_per_pod_queue_size",
-		"inference_objective_running_requests",
-		"inference_pool_ready_pods",
-		"inference_extension_info",
-	}
-	expectedMetrics := make([]string, 0, len(preset)+len(modelServerPods)*numPorts)
-	expectedMetrics = append(expectedMetrics, preset...)
-
-	for _, modelServerPod := range modelServerPods {
-		for rank := range numPorts {
-			metricQueueSize := fmt.Sprintf(
-				"inference_pool_per_pod_queue_size{model_server_pod=\"%s-rank-%d\",name=\"%s\"}",
-				modelServerPod.Name,
-				rank,
-				modelServerName,
-			)
-			expectedMetrics = append(expectedMetrics, metricQueueSize)
-		}
-	}
-
-	gomega.Eventually(func() error {
-		// Execute the metrics scrape command inside the curl pod.
-		resp, err := testutils.ExecCommandInPod(testConfig, "curl", "curl", metricScrapeCmd)
-		if err != nil {
-			return err
-		}
-		// Verify that we got a 200 OK response.
-		if !strings.Contains(resp, "200 OK") {
-			return fmt.Errorf("did not get 200 OK: %s", resp)
-		}
-		// Check if all expected metrics are present in the metrics output.
-		for _, metric := range expectedMetrics {
-			if !strings.Contains(resp, metric) {
-				return fmt.Errorf("expected metric %s not found in metrics output", metric)
-			}
-		}
-		return nil
-	}, testConfig.ReadyTimeout, curlInterval).Should(gomega.Succeed())
-}
-
-func getMetricsReaderToken(k8sClient client.Client) (string, error) {
-	secret := &corev1.Secret{}
-	err := k8sClient.Get(testConfig.Context, types.NamespacedName{Namespace: testConfig.NsName, Name: metricsReaderSecretName}, secret)
-	if err != nil {
-		return "", err
-	}
-	return string(secret.Data["token"]), nil
-}
-
-// findReadyPod finds the first EPP pod that has a "Ready" status condition.
-// It's used to target the leader pod in an HA setup.
-func findReadyPod() *corev1.Pod {
-	var readyPod *corev1.Pod
-	gomega.Eventually(func(g gomega.Gomega) {
-		podList := &corev1.PodList{}
-		err := testConfig.K8sClient.List(testConfig.Context, podList, client.InNamespace(testConfig.NsName), client.MatchingLabels{"app": inferExtName})
-		g.Expect(err).NotTo(gomega.HaveOccurred())
-
-		foundReadyPod := false
-		for i := range podList.Items {
-			pod := &podList.Items[i]
-			for _, cond := range pod.Status.Conditions {
-				if cond.Type == corev1.PodReady && cond.Status == corev1.ConditionTrue {
-					g.Expect(pod.Status.PodIP).NotTo(gomega.BeEmpty(), "Ready pod must have an IP")
-					readyPod = pod
-					foundReadyPod = true
-					break // break inner loop
-				}
-			}
-			if foundReadyPod {
-				break // break outer loop
-			}
-		}
-		g.Expect(foundReadyPod).To(gomega.BeTrue(), "No ready EPP pod found")
-	}, testConfig.ReadyTimeout, testConfig.Interval).Should(gomega.Succeed())
-	return readyPod
-}
-
-// getMetricsScrapeCommand returns the command to scrape the /metrics endpoint.
-func getMetricsScrapeCommand(podIP, token string) []string {
-	return []string{
-		"curl", "-i", "--max-time", strconv.Itoa((int)(6 * curlTimeout.Seconds())),
-		"-H", "Authorization: Bearer " + token, fmt.Sprintf("http://%s:%d/metrics", podIP, 9090),
-	}
-}
-
-// getCurlCommand returns the command, as a slice of strings, for curl'ing
-// the test model server at the given name, namespace, port, and model name.
-// This command gets executed by a dummy pod that communicates with Envoy
-func getCurlCommand(name, ns, port, model string, timeout time.Duration, api string, promptOrMessages any, streaming bool) []string {
-	body := map[string]any{
-		"model":       model,
-		"max_tokens":  100,
-		"temperature": 0,
-	}
-	body["model"] = model
-	switch api {
-	case "/completions":
-		body["prompt"] = promptOrMessages
-	case "/chat/completions":
-		body["messages"] = promptOrMessages
-	case apiEmbeddings:
-		body["input"] = promptOrMessages
-		delete(body, "max_tokens")
-		delete(body, "temperature")
-	}
-	if streaming && api != apiEmbeddings {
-		body["stream"] = true
-		body["stream_options"] = map[string]any{
-			"include_usage": true,
-		}
-	}
-	b, err := json.Marshal(body)
-	gomega.Expect(err).NotTo(gomega.HaveOccurred())
-	return []string{
-		"curl",
-		"-i",
-		"--max-time",
-		strconv.Itoa((int)(timeout.Seconds())),
-		fmt.Sprintf("%s.%s.svc:%s/v1%s", name, ns, port, api),
-		"-H",
-		"Content-Type: application/json",
-		"-H",
-		"Cache-Control: no-cache",
-		"-H",
-		fmt.Sprintf("%v: inferenceobjective-sample", metadata.ObjectiveKey),
-		"-H",
-		fmt.Sprintf("%v: %s", metadata.ModelNameRewriteKey, targetModelName),
-		"-H",
-		"Connection: close",
-		"-d",
-		string(b),
-	}
-}
-
-// buildContainerPorts constructs a slice of corev1.ContainerPort starting from 'start' with 'count' ports.
-func buildContainerPorts(start int, count int) []corev1.ContainerPort {
-	ports := make([]corev1.ContainerPort, count)
-	for i := range count {
-		portNum := int32(start + i)
-		ports[i] = corev1.ContainerPort{
-			Name:          fmt.Sprintf("http-%d", portNum),
-			ContainerPort: portNum,
-			Protocol:      corev1.ProtocolTCP,
-		}
-	}
-	return ports
-}
-
-// buildTargetPorts constructs a slice of v1.Port starting from 'start' with 'count' ports.
-func buildTargetPorts(start int, count int) []v1.Port {
-	ports := make([]v1.Port, count)
-	for i := range count {
-		ports[i] = v1.Port{
-			Number: v1.PortNumber(start + i),
-		}
-	}
-	return ports
-}
-
-// waitForDeploymentRollout waits until the Deployment has completed its update.
-// It ensures that the new version is fully rolled out and available.
-func waitForDeploymentRollout(tc *testutils.TestConfig, deploy *appsv1.Deployment) {
-	ginkgo.By(fmt.Sprintf("Waiting for Deployment %s/%s to complete rollout", deploy.Namespace, deploy.Name))
-
-	key := types.NamespacedName{Name: deploy.Name, Namespace: deploy.Namespace}
-
-	gomega.Eventually(func() error {
-		currentDeploy := &appsv1.Deployment{}
-		if err := tc.K8sClient.Get(tc.Context, key, currentDeploy); err != nil {
-			return err
-		}
-
-		if currentDeploy.Generation > currentDeploy.Status.ObservedGeneration {
-			return errors.New("deployment generation not observed yet")
-		}
-
-		desiredReplicas := *currentDeploy.Spec.Replicas
-
-		if currentDeploy.Status.UpdatedReplicas < desiredReplicas {
-			return fmt.Errorf("waiting for updated replicas: %d/%d", currentDeploy.Status.UpdatedReplicas, desiredReplicas)
-		}
-
-		if currentDeploy.Status.AvailableReplicas < desiredReplicas {
-			return fmt.Errorf("waiting for available replicas: %d/%d", currentDeploy.Status.AvailableReplicas, desiredReplicas)
-		}
-
-		if currentDeploy.Status.Replicas > desiredReplicas {
-			return fmt.Errorf("waiting for old replicas to terminate: %d > %d", currentDeploy.Status.Replicas, desiredReplicas)
-		}
-
-		return nil
-	}, testConfig.ReadyTimeout, testConfig.Interval).Should(gomega.Succeed(), "Deployment failed to roll out within timeout")
-
-	ginkgo.By("Deployment rollout complete")
-}
-
-// waitForDeploymentReady waits for the Deployment to have all replicas ready.
-func waitForDeploymentReady(tc *testutils.TestConfig, deploy *appsv1.Deployment) {
-	ginkgo.By(fmt.Sprintf("waiting for Deployment %s/%s to be ready", deploy.Namespace, deploy.Name))
-
-	key := types.NamespacedName{Name: deploy.Name, Namespace: deploy.Namespace}
-
-	gomega.Eventually(func() error {
-		current := &appsv1.Deployment{}
-		if err := tc.K8sClient.Get(tc.Context, key, current); err != nil {
-			return err
-		}
-
-		if current.Status.Replicas != current.Status.ReadyReplicas {
-			return fmt.Errorf("replicas mismatch: expected %d, got %d ready",
-				current.Status.Replicas, current.Status.ReadyReplicas)
-		}
-
-		if current.Status.ReadyReplicas == 0 {
-			return errors.New("no replicas are ready yet")
-		}
-
-		return nil
-	}, testConfig.ReadyTimeout, testConfig.Interval).Should(gomega.Succeed())
-}
-
-// generateTraffic sends multiple concurrent requests using the provided curl command.
-func generateTraffic(
-	curlCmd []string,
-	batches int,
-	semaphore chan struct{},
-) error {
-	var wg sync.WaitGroup
-	errorCh := make(chan error, batches)
-
-	for i := range batches {
-		wg.Add(1)
-		semaphore <- struct{}{}
-
-		go func(requestNum int) {
-			defer wg.Done()
-			defer func() { <-semaphore }()
-
-			var err error
-			// RETRY LOOP
-			for attempt := 0; attempt <= maxRetries; attempt++ {
-				_, err = testutils.ExecCommandInPod(testConfig, "curl", "curl", curlCmd)
-				if err == nil {
-					return // Success!
-				}
-
-				time.Sleep(backoff)
-			}
-
-			// If we get here, we failed all retries
-			errorCh <- fmt.Errorf("request %d failed: %w", requestNum, err)
-		}(i)
-	}
-
-	wg.Wait()
-	close(errorCh)
-
-	// Collect any errors that occurred
-	failures := make([]error, 0, batches)
-	for err := range errorCh {
-		failures = append(failures, err)
-	}
-
-	if len(failures) > 0 {
-		return fmt.Errorf("found %d failed requests: %v", len(failures), failures)
-	}
-
-	return nil
-}
-
-// getPodsByLabel lists pods in a given namespace that have a specific label key-value pair.
-func getPodsByLabel(ctx context.Context, k8sClient client.Client, namespace, labelKey, labelValue string) ([]corev1.Pod, error) {
-	podList := &corev1.PodList{}
-	labels := map[string]string{labelKey: labelValue}
-
-	listOptions := []client.ListOption{
-		client.InNamespace(namespace),
-		client.MatchingLabels(labels),
-	}
-
-	if err := k8sClient.List(ctx, podList, listOptions...); err != nil {
-		return nil, fmt.Errorf("failed to list pods with label %s=%s in namespace %s: %w", labelKey, labelValue, namespace, err)
-	}
-	return podList.Items, nil
-}
-
-// generateSequence generates a sequence of integers starting from 'start' with 'count' numbers.
-func generateSequence(start int, count int) []int {
-	nums := make([]int, count)
-	for i := range count {
-		nums[i] = start + i
-	}
-	return nums
-}
